+++ drivers/sqlserver/sqlserver.go
@@ -855,8 +855,7 @@ func setIdentityInsert(ctx context.Context, db sqlz.DB, tbl string, on bool) err
 // tblfmt formats a table name for use in a query. The arg can be a string,
 // or a tablefq.T.
 func tblfmt[T string | tablefq.T](tbl T) string {
-	tfq := tablefq.From(tbl)
-	return tfq.Render(stringz.DoubleQuote)
+	return tablefq.Format(tbl, stringz.DoubleQuote)
 }
 
 // genDropSchemaObjectsStmt generates a SQL statement that drops all